+++ labware-library/src/labware-creator/components/HeightGuidingText.js
@@ -1,0 +1,49 @@
+// @flow
+import * as React from 'react'
+import type { LabwareType } from '../fields'
+
+const HeightGuidingText = (props: {| labwareType: ?LabwareType |}) => {
+  const { labwareType } = props
+  const footer = (
+    <p>
+      The height measurement informs the robot of the top and bottom of your
+      labware.
+    </p>
+  )
+  if (labwareType === 'tubeRack') {
+    return (
+      <>
+        <p>Place your tubes inside the rack.</p>
+        <p>
+          Reference{' '}
+          <strong>from the top of the tube to bottom of the rack.</strong>{' '}
+          Include any well lip. Exclude any cover or cap.
+        </p>
+        {footer}
+      </>
+    )
+  }
+  if (labwareType === 'aluminumBlock') {
+    return (
+      <>
+        <p>Put your labware on top of the aluminum block.</p>
+        <p>
+          Reference{' '}
+          <strong>
+            from the top of your labware to the bottom of the block.
+          </strong>{' '}
+          Include any well or tube lip. Exclude any cover or cap.
+        </p>
+        {footer}
+      </>
+    )
+  }
+  return (
+    <>
+      <p>Include any well lip in the measurement. Exclude any cover or cap.</p>
+      {footer}
+    </>
+  )
+}
+
+export default HeightGuidingText
