+++ labware-library/src/labware-creator/components/optionsWithImages/optionsWithImages.css
@@ -1,0 +1,17 @@
+@import '@opentrons/components';
+
+.hidden {
+  display: none;
+}
+
+.radio_image_label {
+  text-anchor: middle;
+  display: flex;
+  align-items: center;
+  padding: 0.25rem 0;
+}
+
+.radio_image {
+  padding: 0 0.5rem;
+  height: 1.25rem;
+}
